Further separation between loading opi and bob files.
